requirements.txt
Django==3.0.7
django-allauth==0.42.0
psycopg2-binary==2.8.5
python-decouple==3.3